feat: handle allowance with authentication
